refactor(add_role_badges): refactor EditAssociation into SaveAssociation for clearability, and to be used in both contexts
